Turn update
You can now select the correct move when clicking on it

After selecting move, it will show on the top of the screen your move, then after selecting enough moves for 6 actions, the turn will run!

Graphics will be done in the future but for now, game logic on selecting moves part completely works! :)))))) happy me :))))