Major Game Logic Update
Only did this for forwards movement but seems to be working really nice.

So, now instead of resolving actions directly, it first checks all actions to see if they are allowed to be resolved during frame = -1, then after checking, frame becomes 0 and the resolve action function is then called to resolve the action (if only the action CAN be resolved).

Also made it so that player will score if the other player moves out of line!