Classes Framework Fully Implemented YAY BIG CHANGES
Ok this is very long but:

Firstly:
Classes introduced for a much much much cleaner implementation of the system! So how it works is that the game now actually uses the Action class, and has child classes for each and every type of action that exists, for example, there is the Hit class for hit, etc... So the Move class now stores the Action class instead of purely the symbol. The Game Engine class now calls back to the Action class's check function to see if the Action can be done, instead of purely doing everything through the Game Engine class. The Action class btw also has different states, now very nicely organised into dicts instead of variables indicating each thing (btw use this for other projects as well! organise state variables into dicts! It is very useful!). So what this does is that it can store each type of state, e.g. is the action scoring? is the action allowing movement? is the action allowing pushing? Then all of these states are changed depending on the game, then the Game Engine class can resolve the actions accordingly.

Secondly:
Game logic is now semi implemented (after the class framework is complete, game logic is semi done because the charge doesn't work on attacks yet, getting pushed off mat doesn't let you score yet (i think? or maybe i already implemented it lol idk i need to check), and you cant physically win the game yet. Though game logic will be worked on right after this commit). Anyway, the game logic is semi complete so the player can actually hit the enemy and score, move to different tiles, and push if needed, etc etc

Thirdly:
ALL_MOVES are now made cooler! So there is a place where you can easily input moves into, then it gets converted to the Action and Move classes right when the game loads. This is good because instead of me having to input every single move by its class, I can literally type a string full of symbols its easier lol

So yea that was a lot of changes! But mainly need to work on game logic next, and then organise the files, then animation :+1: