Big brain score push move functions
This is the biggest brain part in this project so far :O

Basically, I have set up 3 functions, for scoring, pushing, and movement. These functions will do the computation of whether the two actions in same slot in 1 turn should trigger scoring, pushing, or movement for either of the actions.

This was originally done in a massive bunch of if statements, and then one of the other commits, I have changed everything into classes, but its still essentially if statements in a more organised manner. BUT NOW, its in a function, which means I wont miss any cases left unchecked, and no more if statements (there's still some, but no more if statements for each and every case) because its all done in addition / multiplication computations that just checks everything clearly.

wrote down all information on ipad for reference when writing report :)

This implementation actually fixed so many bugs so this is soooo based omg