ok big changes
so I kinda finished the dqn, but not fully

Lemme write this here before i forget tmr:
DO THE EPISODES THING - so basically add the reward to the previous move, if action == -1, because it is the previous move that caused the win / loss if this turn is skipped. (ok this will sound absolutely dumb but as long as i get it is ok becuz its just a note for myself)

anyway highly improved the dqn i hope?

Also made it so that the dqn has a mask for the q values, so only the hand will be considered each turn and changed:
mask is a zeros function, then take the index from the hand, then set them to 1,  then let the q values multiply by this, making anything not in hand = 0

In the predict function, its -inf instead because argmax is taken, just in case theres negative values, 0 isn't a good choice for the predict function.

Also fixed the bug (still fixing tbh because of the episodes thing) where when the turn is skipped, training selects a different move because of -1 as index back then.

Also learning RNN, once i finish with the episodes thing, this i will try do, so rn RNN doesn';t have mask and anything because it was made before I figured out the bug and the stuff on the hand as well.

Also trained a bunch of models that might be potentially useful in the future because some have really high winrates against random bots like 80% even with the bugs persisting. So that is promising and im keeping that.

Also changed the rewards a little: 20 for win, 10 for tie, -10 for loss

Also in dqn added normalisation for q values using min max normalisation to rescale the values in the input so its within [0,1]

Mainly lots of fixes and training, then realisng there should be more fixes, so now i ahve to do it again, committing so  I feel better lmao